Enhance: Rename and update main token ABI json
